Add CompiledOps
